+++ CHANGELOG.md
@@ -7,10 +7,18 @@ and this project adheres to [Dart Package Versioning](https://dart.dev/tools/pub
 
 ## [Unreleased]
 
+### Added
+
+- Len class for field length validation —
+  [34](https://github.com/dartoos-dev/formdator/issues/34).
+
 ### Changed
 
-- make MaxLen and MinLen accept String, Iterable and Map as input value —
-  [36](https://github.com/dartoos-dev/formdator/issues/36).
+- move all functionality from the MaxLen and MinLen classes to Len the class.
+
+### Removed
+
+- MaxLen and MinLen classes — **BREAKING CHANGE**.
 
 ## [0.3.1] - 2021-06-19